+++ CopilotKit/packages/runtime/src/service-adapters/langchain/utils.ts
@@ -51,27 +51,32 @@ export function convertMessageToLangChainMessage(message: Message): BaseMessage 
 export function convertJsonSchemaToZodSchema(jsonSchema: any, required: boolean): z.ZodSchema {
   if (jsonSchema.type === "object") {
     const spec: { [key: string]: z.ZodSchema } = {};
+
+    if (!jsonSchema.properties || !Object.keys(jsonSchema.properties).length) {
+      return !required ? z.object(spec).optional() : z.object(spec);
+    }
+
     for (const [key, value] of Object.entries(jsonSchema.properties)) {
       spec[key] = convertJsonSchemaToZodSchema(
         value,
         jsonSchema.required ? jsonSchema.required.includes(key) : false,
       );
     }
-    let schema = z.object(spec);
-    return !required ? schema.optional() : schema;
+    let schema = z.object(spec).describe(jsonSchema.description);
+    return required ? schema : schema.optional();
   } else if (jsonSchema.type === "string") {
     let schema = z.string().describe(jsonSchema.description);
-    return !required ? schema.optional() : schema;
+    return required ? schema : schema.optional();
   } else if (jsonSchema.type === "number") {
     let schema = z.number().describe(jsonSchema.description);
-    return !required ? schema.optional() : schema;
+    return required ? schema : schema.optional();
   } else if (jsonSchema.type === "boolean") {
     let schema = z.boolean().describe(jsonSchema.description);
-    return !required ? schema.optional() : schema;
+    return required ? schema : schema.optional();
   } else if (jsonSchema.type === "array") {
     let itemSchema = convertJsonSchemaToZodSchema(jsonSchema.items, true);
-    let schema = z.array(itemSchema);
-    return !required ? schema.optional() : schema;
+    let schema = z.array(itemSchema).describe(jsonSchema.description);
+    return required ? schema : schema.optional();
   }
 }
 
@@ -196,6 +201,13 @@ export async function streamLangChainResponse({
 
     let mode: "function" | "message" | null = null;
 
+    const toolCallDetails = {
+      name: null,
+      id: null,
+      index: null,
+      prevIndex: null,
+    };
+
     while (true) {
       try {
         const { done, value } = await reader.read();
@@ -208,10 +220,21 @@ export async function streamLangChainResponse({
 
         if (isAIMessageChunk(value)) {
           let chunk = value.tool_call_chunks?.[0];
-          toolCallName = chunk?.name;
-          toolCallId = chunk?.id;
           toolCallArgs = chunk?.args;
           hasToolCall = chunk != undefined;
+          if (chunk?.name) toolCallDetails.name = chunk.name;
+          // track different index on the same tool cool
+          if (chunk?.index != null) {
+            toolCallDetails.index = chunk.index; // 1
+            if (toolCallDetails.prevIndex == null) toolCallDetails.prevIndex = chunk.index;
+          }
+          // Differentiate when calling the same tool but with different index
+          if (chunk?.id)
+            toolCallDetails.id = chunk.index != null ? `${chunk.id}-idx-${chunk.index}` : chunk.id;
+
+          // Assign to internal variables that the entire script here knows how to work with
+          toolCallName = toolCallDetails.name;
+          toolCallId = toolCallDetails.id;
         } else if (isBaseMessageChunk(value)) {
           let chunk = value.additional_kwargs?.tool_calls?.[0];
           toolCallName = chunk?.function?.name;
@@ -237,7 +260,7 @@ export async function streamLangChainResponse({
 
         // If we send a new message type, send the appropriate start event.
         if (mode === null) {
-          if (hasToolCall) {
+          if (hasToolCall && toolCallId && toolCallName) {
             mode = "function";
             eventStream$.sendActionExecutionStart(toolCallId, toolCallName);
           } else if (content) {
@@ -252,6 +275,12 @@ export async function streamLangChainResponse({
             Array.isArray(content) ? (content[0]?.text ?? "") : content,
           );
         } else if (mode === "function" && toolCallArgs) {
+          // For calls of the same tool with different index, we seal last tool call and register a new one
+          if (toolCallDetails.index !== toolCallDetails.prevIndex) {
+            eventStream$.sendActionExecutionEnd();
+            eventStream$.sendActionExecutionStart(toolCallId, toolCallName);
+            toolCallDetails.prevIndex = toolCallDetails.index;
+          }
           eventStream$.sendActionExecutionArgs(toolCallArgs);
         }
       } catch (error) {