+++ CopilotKit/packages/react-core/CHANGELOG.md
@@ -1,5 +1,42 @@
 # ui
 
+## 1.3.15
+
+### Patch Changes
+
+- pass description for array and object action parameters in langchain adapter
+- Updated dependencies
+  - @copilotkit/runtime-client-gql@1.3.15
+  - @copilotkit/shared@1.3.15
+
+## 1.3.14
+
+### Patch Changes
+
+- Add data-test-id to some elements for testing
+- Updated dependencies
+  - @copilotkit/runtime-client-gql@1.3.14
+  - @copilotkit/shared@1.3.14
+
+## 1.3.13
+
+### Patch Changes
+
+- fix usage of one-at-a-time tool when called multiple times
+- Updated dependencies
+  - @copilotkit/runtime-client-gql@1.3.13
+  - @copilotkit/shared@1.3.13
+
+## 1.3.12
+
+### Patch Changes
+
+- - enable dynamic parameters in langchain adapter tool call
+  - fix unparsable action arguments causing tool call crashes
+- Updated dependencies
+  - @copilotkit/runtime-client-gql@1.3.12
+  - @copilotkit/shared@1.3.12
+
 ## 1.3.11
 
 ### Patch Changes